fix(doom/f_wipe): compute the column-xform copy size in size_t
wipe_shittyColMajorXform does memcpy(array, dest, width*height*2). width and
height are int, so the byte count is computed in 32-bit and only widened to
memcpy's size_t afterwards — it can wrap before the conversion. CodeQL flags it
as cpp/integer-multiplication-cast-to-long (CWE-190).

Cast the first operand to size_t so the product is 64-bit. No behavioural change
at real screen dimensions.

+++ kernel/doom_src/f_wipe.c
@@ -55,7 +55,7 @@ wipe_shittyColMajorXform
 	for(x=0;x<width;x++)
 	    dest[x*height+y] = array[y*width+x];
 
-    memcpy(array, dest, width*height*2);
+    memcpy(array, dest, (size_t)width*height*2);   // 64-bit multiply before widening to size_t
 
     Z_Free(dest);
